refactor(dialog/expansion-panel): optimize CSS
