+++ src/components/dialog/themes/dialog.base.scss
@@ -1,9 +1,14 @@
 @use 'styles/common/component';
 @use 'styles/utilities' as *;
+@use './light/themes' as *;
+
+$theme: $material;
 
 :host {
     border-style: none;
     padding: unset;
+    background: var-get($theme, 'background');
+    border-radius: var-get($theme, 'border-radius');
 }
 
 [part='base'] {
@@ -14,9 +19,13 @@
     border: inherit;
     border-radius: inherit;
     min-width: inherit;
+    box-shadow: var-get($theme, 'elevation');
 }
 
 [part='content'] {
+    color: var-get($theme, 'message-color');
+    padding: rem(14px) rem(24px);
+
     @include type-style('body-1') {
         margin: 0;
     }
@@ -24,6 +33,7 @@
 
 [part='title'],
 [part='footer'] {
+    color: var-get($theme, 'title-color');
     display: flex;
     align-items: center;
 
@@ -32,9 +42,16 @@
     }
 }
 
+[part='title'] {
+    padding-inline: rem(24px);
+    padding-block-start: rem(16px);
+}
+
 [part='footer'] {
     flex-wrap: nowrap;
     justify-content: flex-end;
+    padding: rem(8px);
+    padding-block-start: 0;
 }
 
 ::slotted(*) {
@@ -63,6 +80,7 @@ dialog::backdrop {
     pointer-events: none;
     opacity: 0;
     transition: opacity 350ms ease-out;
+    background: var-get($theme, 'backdrop');
 }
 
 :host([open]) [part='backdrop'] {
