+++ src/components/dialog/themes/shared/dialog.bootstrap.scss
@@ -3,19 +3,9 @@
 
 $theme: $bootstrap;
 
-:host {
-    background: var-get($theme, 'background');
-    border-radius: var-get($theme, 'border-radius');
-}
-
-[part~='backdrop'] {
-    background: var-get($theme, 'backdrop');
-}
-
 [part='base'] {
     min-width: rem(288px);
     border: 1px solid var-get($theme, 'border-color');
-    box-shadow: var-get($theme, 'elevation');
 }
 
 [part='title'],
@@ -24,12 +14,10 @@ $theme: $bootstrap;
         margin: 0;
     }
 
-    color: var-get($theme, 'title-color');
     padding: rem(16px);
 }
 
 [part='content'] {
-    color: var-get($theme, 'message-color');
     padding: rem(16px);
 }
