+++ src/components/dialog/themes/shared/dialog.fluent.scss
@@ -1,35 +1,16 @@
 @use 'styles/utilities' as *;
-@use '../light/themes' as *;
-
-$theme: $fluent;
-
-:host {
-    background: var-get($theme, 'background');
-    border-radius: var-get($theme, 'border-radius');
-}
-
-[part='base'] {
-    box-shadow: var-get($theme, 'elevation');
-}
 
 [part='title'] {
-    color: var-get($theme, 'title-color');
     padding: rem(24px);
     padding-block-start: rem(16px);
 }
 
 [part='content'] {
-    color: var-get($theme, 'message-color');
     padding-inline: rem(24px);
     padding-block-end: rem(20px);
 }
 
 [part='footer'] {
-    color: var-get($theme, 'title-color');
     padding: rem(24px);
     padding-block-start: 0;
-}
-
-[part~='backdrop'] {
-    background: var-get($theme, 'backdrop');
 }
