--- src/components/dialog/themes/shared/dialog.material.scss
@@ -1,34 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../light/themes' as *;
-
-$theme: $material;
-
-:host {
-    background: var-get($theme, 'background');
-    border-radius: var-get($theme, 'border-radius');
-}
-
-[part~='backdrop'] {
-    background: var-get($theme, 'backdrop');
-}
-
-[part='base'] {
-    box-shadow: var-get($theme, 'elevation');
-}
-
-[part='title'] {
-    color: var-get($theme, 'title-color');
-    padding-inline: rem(24px);
-    padding-block-start: rem(16px);
-}
-
-[part='content'] {
-    color: var-get($theme, 'message-color');
-    padding: rem(14px) rem(24px);
-}
-
-[part='footer'] {
-    color: var-get($theme, 'title-color');
-    padding: rem(8px);
-    padding-block-start: 0;
-}
