+++ src/components/expansion-panel/themes/shared/expansion-panel.fluent.scss
@@ -1,38 +1,12 @@
 @use 'styles/utilities' as *;
-@use '../light/themes' as *;
-
-$theme: $fluent;
-
-:host {
-    color: var-get($theme, 'body-color');
-    background: var-get($theme, 'body-background');
-    border-radius: var-get($theme, 'border-radius');
-}
-
-[part='header'] {
-    background: var-get($theme, 'header-background');
-
-    &:focus,
-    &:active {
-        background: var-get($theme, 'header-focus-background');
-    }
-}
-
-[part~='indicator'] {
-    color: var-get($theme, 'header-icon-color');
-}
 
 [part='title'] {
-    color: var-get($theme, 'header-title-color');
-
     &::slotted(*) {
         @include type-style('subtitle-1');
     }
 }
 
 [part='subtitle'] {
-    color: var-get($theme, 'header-description-color');
-
     &::slotted(*) {
         @include type-style('body-2');
     }
@@ -40,12 +14,4 @@ $theme: $fluent;
 
 :host([open]) [part~='content'] {
     @include type-style('caption');
-}
-
-:host([disabled]) {
-    [part~='indicator'],
-    [part='title'],
-    [part='subtitle'] {
-        color: var-get($theme, 'disabled-color');
-    }
 }
