+++ src/components/expansion-panel/themes/shared/expansion-panel.indigo.scss
@@ -1,43 +1,5 @@
 @use 'styles/utilities' as *;
-@use '../light/themes' as *;
-
-$theme: $indigo;
-
-:host {
-    color: var-get($theme, 'body-color');
-    background: var-get($theme, 'body-background');
-    border-radius: var-get($theme, 'border-radius');
-}
-
-[part='header'] {
-    background: var-get($theme, 'header-background');
-
-    &:focus,
-    &:active {
-        background: var-get($theme, 'header-focus-background');
-    }
-}
-
-[part~='indicator'] {
-    color: var-get($theme, 'header-icon-color');
-}
-
-[part='title'] {
-    color: var-get($theme, 'header-title-color');
-}
-
-[part='subtitle'] {
-    color: var-get($theme, 'header-description-color');
-}
 
 :host([open]) [part~='content'] {
     @include type-style('body-1');
-}
-
-:host([disabled]) {
-    [part~='indicator'],
-    [part='title'],
-    [part='subtitle'] {
-        color: var-get($theme, 'disabled-color');
-    }
 }
